Move websms message to be private (internal) class
Add logger to transport
